+++ src/ElementTransformer.php
@@ -2,8 +2,8 @@
 
 namespace craft\elementapi;
 
+use craft\base\Element;
 use craft\base\ElementInterface;
-use craft\helpers\ArrayHelper;
 use League\Fractal\TransformerAbstract;
 
 /**
@@ -16,6 +16,14 @@ class ElementTransformer extends TransformerAbstract
 {
     public function transform(ElementInterface $element): array
     {
-        return ArrayHelper::toArray($element);
+        // Get the serialized custom field values
+        $fields = $element->getSerializedFieldValues();
+
+        // Get the element attributes that aren't custom fields
+        /** @var Element $element */
+        $attributes = array_diff($element->attributes(), array_keys($fields));
+
+        // Return the element as an array merged with its serialized custom field values
+        return array_merge($element->toArray($attributes), $fields);
     }
 }