+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Release Notes for Element API
 
+## Unreleased
+
+### Fixed
+- Fixed a bug where field types that return objects were getting converted to arrays by the default element transformer. Now their [serialized value](https://docs.craftcms.com/api/v3/craft-base-fieldinterface.html#method-serializevalue) is used instead. ([#75](https://github.com/craftcms/element-api/issues/75))
+
 ## 2.5.4 - 2018-07-29
 
 ### Changed